requirements.txt
flask
requests
faiss-cpu
sentence-transformers
numpy

